+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New Features
+
+* Parsing `Defined Names` in xlsx
+
 ## 0.13.0 (2021-05-25)
 
 ### New Features